session radix cache: make the test manual; trim comments
Move test out of test/registered/ (CI globs registered/**) into
test/manual/core/ so it does not run in CI, cut it to the three tests that
matter (shared-prefix, same-leaf last-holder, LRU-neutral), and drop the
verbose docstrings/comments added in the prior commit.
